feat: add AgentCard and TimeCell components for dashboard
- Implemented AgentCard.vue to display agent statistics including rank, wins, podiums, and total time.
- Added computed properties for rank emoji and podium class styling.
- Created TimeCell.vue to represent individual time entries with status indicators and formatted time display.
- Included functionality for running an agent's time and displaying success/error states with appropriate icons.

+++ agents/codex/c/day01/part1.c
@@ -19,14 +19,14 @@ int main(void) {
 
     while (*p) {
         char dir = *p++;
-        long long dist = 0;
+        int dist = 0;
         while (*p >= '0' && *p <= '9') {
             dist = dist * 10 + (*p - '0');
             p++;
         }
         while (*p == '\n' || *p == '\r') p++;
 
-        int step = (int)(dist % 100);
+        int step = dist % 100;
         if (dir == 'R') {
             pos += step;
             if (pos >= 100) pos -= 100;
